refactor | use tsconfig path
