+++ src/pages/join/1/index.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/router';
 import { Button } from 'primereact/button';
 import Navigation from '@/components/Navigation/Navigation';
 import MainLayout from '@/layouts/MainLayout';
-import featureFlag from '@/src/config/featureFlags';
+import featureFlag from '@/config/featureFlags';
 import locale from '@/locales/en.json'; // REFACTOR: read based on language
 
 // REFACTOR: move into a helper function
